# Use local memory in band matrix solve