This PR adds the use of local memory in the band matrix solve.

Closes https://github.com/CliMA/ClimaCore.jl/issues/1738.